+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 0.0.10 (2018-03-30)
 ~~~~~~~~~~~~~~~~~~~
 * Fixed issue where ``-m``-defined macros didn't override ``capture`` macros.
-
+* Fixed issue where ``~`` and environment variables are not expanded in some paths.
 
 0.0.9 (2018-02-03)
 ~~~~~~~~~~~~~~~~~~